% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hist.R
\name{create_hist}
\alias{create_hist}
\title{Create interactive histogram}
\usage{
create_hist(x)
}
\arguments{
\item{x}{Variable from dataset in vector form.}
}
\value{
Plotly figure containing spike histogram (numeric) or traditional histogram (character).
}
\description{
Code adapted from Hmisc package's spike histograms.
}
\examples{
adsl <- safetyData::adam_adsl
create_hist(adsl$AGE)
create_hist(adsl$SITEID)

}
