% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describe.Date}
\alias{describe.character}
\alias{describe.data.frame}
\alias{describe.default}
\alias{describe.factor}
\alias{describe.logical}
\alias{describe.numeric}
\title{Describe atomic vectors or data.frames.}
\usage{
describe(.x)

\method{describe}{numeric}(.x)

\method{describe}{character}(.x)

\method{describe}{factor}(.x)

\method{describe}{Date}(.x)

\method{describe}{logical}(.x)

\method{describe}{data.frame}(.x)

\method{describe}{default}(.x)
}
\arguments{
\item{.x}{a vector or data.frame to be described.}
}
\value{
a data.frame containing important descriptive statistics.
}
\description{
\code{describe} takes vectors and data.frames and returns a data.frame containing important descriptive statistics.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Method for numeric vectors.

\item \code{character}: Method for character vectors.

\item \code{factor}: Method for factors.

\item \code{Date}: Method for Date vectors.

\item \code{logical}: Method for logical vectors.

\item \code{data.frame}: Method for data.frames.

\item \code{default}: Method for default.
}}
\examples{
# Let's make a comment
describe(mtcars)
}

