% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data.R
\name{describe_data}
\alias{describe_data}
\title{Describe data}
\usage{
describe_data(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A tibble containing variable metadata.
}
\description{
Create comprehensive tibble of variable metadata using Hmisc::describe as engine.
}
\examples{

describe_data(safetyData::adam_adsl)

}
