% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data_by.R
\name{describe_data_by}
\alias{describe_data_by}
\title{Describe data by a grouping variable}
\usage{
describe_data_by(data, by)
}
\arguments{
\item{data}{A data frame.}

\item{by}{Name of by variable.}
}
\value{
A tibble containing variable metadata with 1 row per group.
}
\description{
Create comprehensive tibble of variable metadata using Hmisc::describe as engine.
}
\examples{

describe_data_by(safetyData::adam_adsl, by = ARM)

}
