% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_nonnumeric}
\alias{describe_nonnumeric}
\title{Describe non-numeric atomic vectors.}
\usage{
describe_nonnumeric(.x)
}
\arguments{
\item{.x}{a non-numeric atomic vector to be described.}
}
\value{
a data.frame containing important descriptive statistics.
}
\description{
\code{describe_nonnumeric} takes non-numeric atomic vectors and returns a data.frame containing important descriptive statistics.
}
\examples{
# Example
describe_nonnumeric(letters)
}

