% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_numeric}
\alias{describe_numeric}
\title{Describe numeric atomic vectors.}
\usage{
describe_numeric(.x)
}
\arguments{
\item{.x}{a numeric atomic vector or data.frame to be described.}
}
\value{
a data.frame containing important descriptive statistics.
}
\description{
\code{describe_numeric} takes numeric vectors and returns a data.frame containing important descriptive statistics.
}
\examples{
# Example
describe_numeric(1:100)
}

