% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describer.R
\docType{package}
\name{describer}
\alias{describer}
\alias{describer-package}
\title{describer: Describe Data in R Using Common Descriptive Statistics}
\description{
describer: Describe Data in R Using Common Descriptive Statistics
}

