% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describers.R
\name{meansd}
\alias{meansd}
\title{A collection of functions to summarize vectors in matrix format}
\usage{
meansd(var, na.rm = TRUE, varname = NULL, digits = 2)
}
\arguments{
\item{var}{A character or numeric vector}

\item{na.rm}{(default TRUE) whether to remove NA values when summarizing
numeric values}

\item{varname}{An optional character string for the name of the vector
(defaults to object name)}

\item{digits}{The number of significant digits to show}

\item{var2}{A second categorical variable for crosstabulation}

\item{outcome}{The comparison variable, divides the data into categories}
}
\description{
These functions work on character, numeric, or factor
vectors. They are meant to make outputs easier to paste into table.
}
