% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{pd_describe}
\alias{pd_describe}
\title{Describe data.frames with pandas results.}
\usage{
pd_describe(.x)
}
\arguments{
\item{.x}{a vector or data.frame to be described.}
}
\value{
a data.frame containing important descriptive statistics.
}
\description{
\code{pd_describe} takes vectors and data.frames and returns a data.frame containing important descriptive statistics.
}
\examples{
# Example
pd_describe(mtcars)
pd_describe(iris)
}

