% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describe.character}
\alias{describe.data.frame}
\alias{describe.default}
\alias{describe.numeric}
\title{Describe atomic vectors or data.frames.}
\usage{
describe(.x)

\method{describe}{numeric}(.x)

\method{describe}{character}(.x)

\method{describe}{data.frame}(.x)

\method{describe}{default}(.x)
}
\arguments{
\item{.x}{a vector or data.frame to be described.}
}
\value{
a data.frame containing important descriptive statistics.
}
\description{
\code{describe} takes vectors and data.frames and returns a data.frame containing important descriptive statistics.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Method for numeric.

\item \code{character}: Method for character.

\item \code{data.frame}: Method for data.frames.

\item \code{default}: Method for default.
}}
\examples{
# Example
describe(mtcars)
describe(iris)
describe(state.name)
}

