% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describer.R
\name{describer}
\alias{describer}
\title{Create interactive table using Hmisc::describe + reactable}
\usage{
describer(data, data_shared = NULL)
}
\arguments{
\item{data}{Output of `describe_data()`}

\item{data_shared}{[Optional] `data` converted to a `SharedData` object using crosstalk, for use with linked widgets.}
}
\value{
Reactable display
}
\description{
Create interactive table using Hmisc::describe + reactable
}
\examples{

adsl <- safetyData::adam_adsl
adsl_describe <- describe_data(adsl)
describer(adsl_describe)

}
