% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describers.R
\name{describers}
\alias{describers}
\alias{describers}
\alias{meansd}
\alias{describers}
\alias{proportion}
\alias{describers}
\alias{crosstab}
\alias{describers}
\alias{groupMeans}
\alias{describers}
\alias{describe}
\title{A collection of functions to summarize vectors in matrix format}
\usage{
meansd(var, na.rm = TRUE, varname = NULL, digits = 2)

proportion(var, digits = 1)

crosstab(var, var2, digits = 2)

groupMeans(var, outcome, digits = 2)

describe(..., outcome, data, headerRow = NULL, headerFrame = NULL,
  includerHeader = TRUE, deparse.level = 2, digits = 2)
}
\arguments{
\item{var}{A character or numeric vector}

\item{na.rm}{(default TRUE) whether to remove NA values when summarizing
numeric values}

\item{varname}{An optional character string for the name of the vector
(defaults to object name)}

\item{digits}{The number of significant digits to show}

\item{var2}{A second categorical variable for crosstabulation}

\item{outcome}{The comparison variable, divides the data into categories}

\item{...}{Variables to describe by the outcome, can be verctors or
unquoted names when the data argument is supplied}

\item{data}{A data.frame containing the comparison variables}

\item{headerRow}{A character vector for each variable in \ldots that serves
as the descriptive name in the table}

\item{headerFrame}{A data.frame that relates or maps each variable name
to its respective descriptive name}

\item{includeHeaders}{logical (default TRUE) indicating whether or not
to add a header row to the table with group counts for the outcome variable}
}
\description{
These functions work on character, numeric, or factor
vectors. They are meant to make outputs easier to paste into table.
}
\examples{

data(mtcars)

meansd(mtcars$mpg)

with(mtcars, groupMeans(mpg, am))

am <- plyr::mapvalues(mtcars$am, c(0, 1), c("Auto", "Manual"))
groupMeans(mtcars$mpg, am)

}
